feat(redis): share Effect Redis client between Fly and Railway
Pull RESP + makeRead/makeWrite/makeReadWrite into alchemy/Redis so Fly
and Railway bindings do not duplicate the runtime client.
