+++ packages/alchemy/src/Fly/RedisBinding.ts
@@ -2,15 +2,10 @@ import * as Config from "effect/Config";
 import * as Effect from "effect/Effect";
 import * as Redacted from "effect/Redacted";
 import * as Binding from "../Binding.ts";
+import type { Url } from "../Redis/index.ts";
 import type { Resource } from "../Resource.ts";
-import type { RuntimeContext } from "../RuntimeContext.ts";
 import type { ServiceBinding } from "./MountVolume.ts";
-import {
-  REDIS_URL_ENV,
-  RedisCommandError,
-  RedisUrlMissing,
-  type Redis,
-} from "./Redis.ts";
+import { REDIS_URL_ENV, RedisUrlMissing, type Redis } from "./Redis.ts";
 
 /**
  * Shared scaffolding for Upstash Redis bindings.
@@ -19,6 +14,9 @@ import {
  * Deploy-time registers the Redis add-on on the host so Service reconcile
  * writes `REDIS_URL`. Runtime commands use that URL internally — callers
  * never read `Config.redacted`.
+ *
+ * The RESP client lives in `alchemy/Redis`. This file only wires the
+ * Fly host binding.
  *
  * NOT exported from `index.ts`.
  */
@@ -37,106 +35,12 @@ const asPlain = (value: unknown): string | undefined => {
   return undefined;
 };
 
-const readValue = (value: unknown): Effect.Effect<string | undefined> =>
-  Effect.gen(function* () {
-    const direct = asPlain(value);
-    if (direct !== undefined) return direct;
-    if (Effect.isEffect(value)) {
-      return asPlain(yield* value as Effect.Effect<unknown>);
-    }
-    return undefined;
-  });
-
 const redisUrlFromEnv = Config.redacted(REDIS_URL_ENV).pipe(
   Effect.map((value) => Redacted.value(value)),
 );
 
-const encodeBulk = (value: string) => `$${value.length}\r\n${value}\r\n`;
-
-const encodeCommand = (command: string, args: readonly string[]) => {
-  const parts = [command, ...args];
-  return `*${parts.length}\r\n${parts.map(encodeBulk).join("")}`;
-};
-
-const decodeReply = (raw: string): unknown => {
-  if (raw.startsWith("-")) {
-    throw new Error(raw.slice(1).split("\r\n")[0] ?? raw);
-  }
-  if (raw.startsWith("+")) {
-    return raw.slice(1).split("\r\n")[0] ?? "";
-  }
-  if (raw.startsWith(":")) {
-    return Number(raw.slice(1).split("\r\n")[0]);
-  }
-  if (raw.startsWith("$-1")) {
-    return null;
-  }
-  if (raw.startsWith("$")) {
-    const newline = raw.indexOf("\r\n");
-    if (newline < 0) return raw;
-    const body = raw.slice(newline + 2);
-    return body.endsWith("\r\n") ? body.slice(0, -2) : body;
-  }
-  const trimmed = raw.trim();
-  return trimmed.length === 0 ? null : trimmed;
-};
-
-const sendCommand = (
-  url: string,
-  command: string,
-  args: readonly string[],
-): Effect.Effect<unknown, RedisCommandError, RuntimeContext> =>
-  Effect.tryPromise({
-    try: async () => {
-      const Client = (
-        globalThis as {
-          Bun?: {
-            RedisClient?: new (url: string) => {
-              send: (
-                command: string,
-                args: readonly string[],
-              ) => Promise<unknown>;
-            };
-          };
-        }
-      ).Bun?.RedisClient;
-      if (Client !== undefined) {
-        const client = new Client(url);
-        return await client.send(command, [...args]);
-      }
-      const parsed = new URL(url);
-      const tls = parsed.protocol === "rediss:";
-      const socket = await Bun.connect({
-        hostname: parsed.hostname,
-        port: Number(parsed.port || (tls ? 6380 : 6379)),
-        tls,
-      });
-      const user = decodeURIComponent(parsed.username || "default");
-      const pass = decodeURIComponent(parsed.password);
-      const frames: string[] = [];
-      if (pass.length > 0) {
-        frames.push(encodeCommand("AUTH", [user, pass]));
-      }
-      frames.push(encodeCommand(command, args));
-      socket.write(frames.join(""));
-      const reader = socket.readable.getReader();
-      const { value } = await reader.read();
-      reader.releaseLock();
-      socket.end();
-      const raw = new TextDecoder().decode(value ?? new Uint8Array());
-      const replies = raw
-        .split(/(?=[+*:$-])/)
-        .filter((part) => part.length > 0);
-      const reply = pass.length > 0 ? replies.at(-1) : replies[0];
-      return decodeReply(reply ?? raw);
-    },
-    catch: (cause) => new RedisCommandError({ command, cause }),
-  });
-
 export const makeRedisBinding = <Client>(options: {
-  makeClient: (
-    url: Effect.Effect<string, RedisUrlMissing, RuntimeContext>,
-  ) => Client;
+  makeClient: (url: Url) => Client;
 }) =>
   Effect.succeed(
     Effect.fn(function* (redis: Redis) {
@@ -160,17 +64,3 @@ export const makeRedisBinding = <Client>(options: {
       return options.makeClient(url);
     }),
   );
-
-export const redisCommand = (
-  url: Effect.Effect<string, RedisUrlMissing, RuntimeContext>,
-  command: string,
-  args: readonly string[] = [],
-): Effect.Effect<
-  unknown,
-  RedisCommandError | RedisUrlMissing,
-  RuntimeContext
-> =>
-  Effect.gen(function* () {
-    const resolved = yield* url;
-    return yield* sendCommand(resolved, command, args);
-  });
