+++ packages/alchemy/src/Railway/Redis.ts
@@ -16,7 +16,6 @@ import { Unowned } from "../AdoptPolicy.ts";
 import { isResolved } from "../Diff.ts";
 import * as Provider from "../Provider.ts";
 import { Resource } from "../Resource.ts";
-import type { RuntimeContext } from "../RuntimeContext.ts";
 import { createRailwayName, matchesAlchemyPhysicalName } from "./Metadata.ts";
 import { listOwnedProjects, type Project } from "./Project.ts";
 import type { Providers } from "./Providers.ts";
@@ -244,18 +243,10 @@ export const Redis: typeof RedisResource = Object.assign(
   RedisResource,
 );
 
-export class RedisUrlMissing extends Data.TaggedError(
-  "Railway.RedisUrlMissing",
-)<{
-  name: string;
-}> {}
-
-export class RedisCommandError extends Data.TaggedError(
-  "Railway.RedisCommandError",
-)<{
-  command: string;
-  cause: unknown;
-}> {}
+export {
+  CommandError as RedisCommandError,
+  UrlMissing as RedisUrlMissing,
+} from "../Redis/index.ts";
 
 export class RedisNotCreated extends Data.TaggedError(
   "Railway.RedisNotCreated",
@@ -563,119 +554,11 @@ const toAttrs = (input: {
   deploymentStatus: input.instance?.latestDeployment?.status,
 });
 
-const encodeBulk = (value: string) => `$${value.length}\r\n${value}\r\n`;
-
-const encodeCommand = (command: string, args: readonly string[]) => {
-  const parts = [command, ...args];
-  return `*${parts.length}\r\n${parts.map(encodeBulk).join("")}`;
-};
-
-const decodeReply = (raw: string): unknown => {
-  if (raw.startsWith("-")) {
-    throw new Error(raw.slice(1).split("\r\n")[0] ?? raw);
-  }
-  if (raw.startsWith("+")) {
-    return raw.slice(1).split("\r\n")[0] ?? "";
-  }
-  if (raw.startsWith(":")) {
-    return Number(raw.slice(1).split("\r\n")[0]);
-  }
-  if (raw.startsWith("$-1")) {
-    return null;
-  }
-  if (raw.startsWith("$")) {
-    const newline = raw.indexOf("\r\n");
-    if (newline < 0) return raw;
-    const body = raw.slice(newline + 2);
-    return body.endsWith("\r\n") ? body.slice(0, -2) : body;
-  }
-  const trimmed = raw.trim();
-  return trimmed.length === 0 ? null : trimmed;
-};
-
-/**
- * Build a `redis://` URL. Encodes the password. Used by laptop tests
- * against a {@link TcpProxy} and by the runtime client.
- */
-export const redisConnectionUrl = (input: {
-  host: string;
-  port: number;
-  password: string;
-  username?: string;
-}): string => {
-  const username = input.username ?? "default";
-  return `redis://${username}:${encodeURIComponent(input.password)}@${input.host}:${input.port}`;
-};
-
-/**
- * Send one Redis command over RESP. Used by the Read/Write clients and
- * by laptop tests against a public TcpProxy URL.
- */
-export const runRedisCommand = (
-  url: string,
-  command: string,
-  args: readonly string[] = [],
-): Effect.Effect<unknown, RedisCommandError> =>
-  Effect.tryPromise({
-    try: async () => {
-      const Client = (
-        globalThis as {
-          Bun?: {
-            RedisClient?: new (url: string) => {
-              send: (
-                command: string,
-                args: readonly string[],
-              ) => Promise<unknown>;
-            };
-          };
-        }
-      ).Bun?.RedisClient;
-      if (Client !== undefined) {
-        const client = new Client(url);
-        return await client.send(command, [...args]);
-      }
-      const parsed = new URL(url);
-      const tls = parsed.protocol === "rediss:";
-      const socket = await Bun.connect({
-        hostname: parsed.hostname,
-        port: Number(parsed.port || (tls ? 6380 : REDIS_PORT)),
-        tls,
-      });
-      const user = decodeURIComponent(parsed.username || "default");
-      const pass = decodeURIComponent(parsed.password);
-      const frames: string[] = [];
-      if (pass.length > 0) {
-        frames.push(encodeCommand("AUTH", [user, pass]));
-      }
-      frames.push(encodeCommand(command, args));
-      socket.write(frames.join(""));
-      const reader = socket.readable.getReader();
-      const { value } = await reader.read();
-      reader.releaseLock();
-      socket.end();
-      const raw = new TextDecoder().decode(value ?? new Uint8Array());
-      const replies = raw
-        .split(/(?=[+*:$-])/)
-        .filter((part) => part.length > 0);
-      const reply = pass.length > 0 ? replies.at(-1) : replies[0];
-      return decodeReply(reply ?? raw);
-    },
-    catch: (cause) => new RedisCommandError({ command, cause }),
-  });
-
-export const redisCommand = (
-  url: Effect.Effect<string, RedisUrlMissing, RuntimeContext>,
-  command: string,
-  args: readonly string[] = [],
-): Effect.Effect<
-  unknown,
-  RedisCommandError | RedisUrlMissing,
-  RuntimeContext
-> =>
-  Effect.gen(function* () {
-    const resolved = yield* url;
-    return yield* runRedisCommand(resolved, command, args);
-  });
+export {
+  command as redisCommand,
+  connectionUrl as redisConnectionUrl,
+  run as runRedisCommand,
+} from "../Redis/index.ts";
 
 export const RedisProvider = () =>
   Provider.succeed(Redis, {
